requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
Flask==1.0.2
Flask-Login==0.4.1
Flask-OAuthlib==0.9.5
Flask-WTF==0.14.2
google-auth==1.6.1
google-auth-oauthlib==0.2.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mypy==0.641
mypy-extensions==0.4.1
oauthlib==2.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pylint==2.2.2
requests==2.21.0
requests-oauthlib==1.0.0
rope==0.11.0
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.14
SQLAlchemy-Utils==0.33.9
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
